### typeahead: Precompile regexes for removing diacritics.

Precompiling regexes gives a performance increase of around 10-15%
based on tests. See https://jsperf.com/typeahead-regex. This stacks
up when we have a lot of users in an organisation.

This concerns #13192.

**Notes**

On manual testing 4 runs of each type, profiling before and after reveals that we shave off close to 150ms on my machine (before: 480ms, after: 320ms) for the first keypress.

![profile-regex-compare](https://user-images.githubusercontent.com/8033238/65221645-11a17e00-dadb-11e9-89c2-1c0b6b47d55f.png)

### typeahead: Cache subsets of items to match.

Below are the test results for subsequent key presses for both, timelines are at the same scale, and we can see that the spikes are of lesser duration in the second case. Moreover, for the third and fourth letters, we do not observe significant loss of framerate.

Test data: 15000 users, their full_names hashed to md5 to create a varied dataset.

![image](https://user-images.githubusercontent.com/8033238/65240670-7078ee80-daff-11e9-8377-ea84417176f4.png)

**Calls to `matcher()`**

| Token | Before | After |
|:----- |-------:| -----:|
| 2     | 15015  | 15015 |
| 23    | 15015  | 13845 |
| 234   | 15015  |  2162 |
| 2345  | 15015  |   142 |
| 234   | 15015  | 15015 |
| 23    | 15015  | 15015 |
| 2     | 15015  | 15015 |

![image](https://user-images.githubusercontent.com/8033238/65236571-859d4f80-daf6-11e9-8a98-9073bf13d55b.png)