@timabbott 2019-09-19
Nice!  The first commit looks great; I merged it after renaming `regexes` to `diacritic_regexes`.