@timabbott 2019-09-19
It's definitely worth trying that approach instead and seeing how it performs.

I posted a longer comment on the caching commit; hopefully it mailed you because GitHub seems to have disappeared it.