@timabbott 2019-09-19
One other thing I notice from those profiles: I see multiple calls to `query_matches_person`; I wonder if that's because we have the `sender:` and `pm-with:` checks for matches to a person if you just type a letter (without an operator); potentially, reorganizing things to reuse the results there (which should always be identical) could buy us a factor of 2.