@timabbott 2023-10-03
Closing as https://github.com/zulip/zulip/pull/13561 and https://github.com/zulip/zulip/pull/13485 resolved this issue, thanks for all the investigation into this @aero31aero!